+++ src/app/views/pages/academice/subject/subject.component.ts
@@ -23,7 +23,7 @@ export class SubjectComponent implements OnInit {
   classs = {
     ClassID: null
   }
-
+  isOnView = false;
   subject = {
     SubjectName: null,
     SubjectCode: null,
@@ -35,7 +35,8 @@ export class SubjectComponent implements OnInit {
     SubjectName: null,
     SubjectCode: null,
     FK_ClassID: null,
-    SubjectID: null
+    SubjectID: null,
+    IsActive: true
 
   }
 
@@ -48,8 +49,9 @@ export class SubjectComponent implements OnInit {
   ngOnInit() {
     this.addSubjectForm = this.formBuilder.group({
       name: new FormControl(name, Validators.required),
-      subjectcode: new FormControl(name, [Validators.required, Validators.minLength(4)]),
+      SubjectCode: new FormControl(name, [Validators.required, Validators.minLength(4)]),
       class: new FormControl(name, Validators.required),
+      IsActive: new FormControl(name)
     });
 
     this.getAllClass();
@@ -72,15 +74,18 @@ export class SubjectComponent implements OnInit {
 
   //View Subject
   ViewRecord(id) {
+    this.isOnView = false;
+    this.isOnEdit = false;
     this.showModal();
     let c;
+    this.isOnView = true;
     c = this.allSubject.filter(x => x.SubjectID == id);
     console.log("Seleted row data ", c);
     if (c != undefined || c != null) {
-      console.log("Field 1", c[0].FK_ClassID)
+      console.log("Field 1", c[0].ClassNo)
       console.log("Field 1", c[0].SubjectCode)
       console.log("Field 1", c[0].SubjectName)
-      this.addSubjectForm.setValue({ name: c[0].SubjectName, subjectcode: c[0].SubjectCode, class: c[0].FK_ClassID })
+      this.addSubjectForm.setValue({ name: c[0].SubjectName, SubjectCode: c[0].SubjectCode, class: c[0].ClassNo, IsActive: c[0].IsActive })
       this.addSubjectForm.disable();
     }
   }
@@ -91,18 +96,19 @@ export class SubjectComponent implements OnInit {
   //Delete Subject
   deleteRecord(id) {
     console.log("Subject id to delete", id)
+    if (confirm("Are you sure ?")) {
+      this.DeleteSubject.SubjectID = id;
+      if (id != null || id != undefined) {
 
-    this.DeleteSubject.SubjectID = id;
-    if (id != null || id != undefined) {
+        this.apiService.subjectService.deleteSubject(this.DeleteSubject).subscribe((res: any) => {
+          this.getallSubject();
+          this.notification.create("success", "Deleted", "Subject Deleted Successfully");
 
-      this.apiService.subjectService.deleteSubject(this.DeleteSubject).subscribe((res: any) => {
-        this.getallSubject();
-        this.notification.create("success", "Deleted", "Subject Deleted Successfully");
+        }, (err) => {
+          this.notification.create("error", "Failed", "Subjcet Deleting Failed")
 
-      }, (err) => {
-        this.notification.create("error", "Failed", "Subjcet Deleting Failed")
-
-      })
+        })
+      }
     }
   }
 
@@ -121,6 +127,7 @@ export class SubjectComponent implements OnInit {
   editRecord(id) {
     this.addSubjectForm.enable();
     this.isOnEdit = true;
+    this.isOnView = false;
     this.showModal();
     let c;
     c = this.allSubject.filter(x => x.SubjectID == id);
@@ -128,9 +135,9 @@ export class SubjectComponent implements OnInit {
     if (c != undefined || c != null) {
       console.log("Field 1", c[0].SubjectCode)
       console.log("Field 1", c[0].SubjectName)
-      console.log("Field 1", c[0].FK_ClassID)
+      console.log("Field 1", c[0].ClassNo)
       this.SubjectID = c[0].SubjectID;
-      this.addSubjectForm.setValue({ name: c[0].SubjectName, subjectcode: c[0].SubjectCode, class: c[0].FK_ClassID })
+      this.addSubjectForm.setValue({ name: c[0].SubjectName, SubjectCode: c[0].SubjectCode, class: c[0].ClassNo, IsActive: c[0].IsActive })
     }
   }
 
@@ -140,9 +147,10 @@ export class SubjectComponent implements OnInit {
     if (this.addSubjectForm.invalid) {
       return;
     } else {
-      this.subject.SubjectCode = this.addSubjectForm.get('subjectcode').value;
+      this.subject.SubjectCode = this.addSubjectForm.get('SubjectCode').value;
       this.subject.SubjectName = this.addSubjectForm.get('name').value;
       this.subject.FK_ClassID = this.addSubjectForm.get('class').value;
+      this.subject.IsActive = this.addSubjectForm.get('IsActive').value;
       console.log("Form data", this.addSubjectForm.value);
       if (!this.isOnEdit) {
         this.apiService.subjectService.createSubject(this.subject).subscribe((res: any) => {
@@ -152,7 +160,6 @@ export class SubjectComponent implements OnInit {
           console.log("Data before isnerting", this.subject)
           this.notification.create("success", "Success", "Subject Added Successfully");
         }, (err) => {
-
           console.log(err)
           this.addSubjectForm.reset();
           this.isVisible = false;
@@ -160,9 +167,10 @@ export class SubjectComponent implements OnInit {
         })
       }
       else {
-        this.updateSubject.SubjectCode = this.addSubjectForm.get('subjectcode').value;
+        this.updateSubject.SubjectCode = this.addSubjectForm.get('SubjectCode').value;
         this.updateSubject.SubjectName = this.addSubjectForm.get('name').value;
         this.updateSubject.FK_ClassID = this.addSubjectForm.get('class').value;
+        this.updateSubject.IsActive = this.addSubjectForm.get('IsActive').value;
         this.updateSubject.SubjectID = this.SubjectID;
         this.apiService.subjectService.updateSubject(this.updateSubject).subscribe((res: any) => {
           this.isVisible = false;
@@ -189,6 +197,11 @@ export class SubjectComponent implements OnInit {
 
   showModal(): void {
     this.isVisible = true;
+  }
+  showAddModal(){
+    this.isOnView = false;
+    this.isOnEdit = false;
+    this.showModal();
   }
   hideModal() {
 
@@ -264,15 +277,15 @@ export class SubjectComponent implements OnInit {
       ]
     });
 
-   // tslint:disable-next-line:max-line-length
-   $('div.dt-buttons button:nth-child(1)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-warning').append('&nbsp;&nbsp;<i class="fa fa-table"> </i>');
-   // tslint:disable-next-line:max-line-length
-   $('div.dt-buttons button:nth-child(2)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-success').append('&nbsp;&nbsp;<i class="fa fa-columns"> </i>');
-   // tslint:disable-next-line:max-line-length
-   $('div.dt-buttons button:nth-child(3)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-info').append('&nbsp;&nbsp;<i class="fa fa-file"> </i>');
-   // tslint:disable-next-line:max-line-length
-   $('div.dt-buttons button:nth-child(4)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-danger').append('&nbsp;&nbsp;<i class="fa fa-print"> </i>');
-   // $('div.dt-buttons button:nth-child(5)').removeClass('dt-button buttons-copy buttons-html5')
+    // tslint:disable-next-line:max-line-length
+    $('div.dt-buttons button:nth-child(1)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-warning').append('&nbsp;&nbsp;<i class="fa fa-table"> </i>');
+    // tslint:disable-next-line:max-line-length
+    $('div.dt-buttons button:nth-child(2)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-success').append('&nbsp;&nbsp;<i class="fa fa-columns"> </i>');
+    // tslint:disable-next-line:max-line-length
+    $('div.dt-buttons button:nth-child(3)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-info').append('&nbsp;&nbsp;<i class="fa fa-file"> </i>');
+    // tslint:disable-next-line:max-line-length
+    $('div.dt-buttons button:nth-child(4)').removeClass('dt-button buttons-copy buttons-html5').addClass('btn btn-outline-danger').append('&nbsp;&nbsp;<i class="fa fa-print"> </i>');
+    // $('div.dt-buttons button:nth-child(5)').removeClass('dt-button buttons-copy buttons-html5')
     //   .addClass('btn btn-outline-danger').append('<i class="fa fa-save"> </>');
     $('div.dt-buttons span').addClass('text');
